requirements.txt
pdf2image==1.17.0
pytesseract==0.3.10
Pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.24.3
langdetect==1.0.9
nltk==3.8.1
spacy==3.7.2
openai==1.88.0
python-dotenv==1.0.0
tqdm==4.66.1
pandas==2.0.3